Move files to Integration folder
