+++ api/app/tests/Integration/Controller-Ndr/AccountControllerTest.php
@@ -2,7 +2,7 @@
 
 namespace App\Tests\Unit\Controller\Ndr;
 
-use App\Tests\Unit\Controller\AbstractTestController;
+use app\tests\Integration\Controller\AbstractTestController;
 
 class AccountControllerTest extends AbstractTestController
 {
