+++ api/app/tests/Integration/Controller/AbstractTestController.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace App\Tests\Unit\Controller;
+namespace app\tests\Integration\Controller;
 
 use App\Service\BruteForce\AttemptsIncrementalWaitingChecker;
 use App\Service\BruteForce\AttemptsInTimeChecker;
 use App\Service\JWT\JWTService;
-use App\Tests\Unit\Fixtures;
+use app\tests\Integration\Fixtures;
 use Doctrine\ORM\EntityManager;
 use Symfony\Bundle\FrameworkBundle\Client;
 use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
