+++ api/app/tests/Integration/Controller/AuthControllerTest.php
@@ -5,6 +5,7 @@ namespace App\Tests\Unit\Controller;
 use App\Entity\User;
 use App\Service\BruteForce\AttemptsIncrementalWaitingChecker;
 use App\Service\BruteForce\AttemptsInTimeChecker;
+use app\tests\Integration\Controller\AbstractTestController;
 
 class AuthControllerTest extends AbstractTestController
 {
