+++ api/app/tests/Integration/Controller/ChecklistControllerTest.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace App\Tests\Unit\Controller;
 
 use App\Entity\Report\Checklist;
+use app\tests\Integration\Controller\AbstractTestController;
 
 class ChecklistControllerTest extends AbstractTestController
 {
