+++ api/app/tests/Integration/Controller/ClientControllerTest.php
@@ -4,6 +4,7 @@ namespace App\Tests\Unit\Controller;
 
 use App\Entity\Client;
 use App\Entity\Ndr\Ndr;
+use app\tests\Integration\Controller\AbstractTestController;
 
 class ClientControllerTest extends AbstractTestController
 {
