+++ api/app/tests/Integration/Controller/CoDeputyControllerTest.php
@@ -2,7 +2,7 @@
 
 namespace App\Controller;
 
-use App\Tests\Unit\Controller\AbstractTestController;
+use app\tests\Integration\Controller\AbstractTestController;
 
 class CoDeputyControllerTest extends AbstractTestController
 {
