+++ api/app/tests/Integration/Controller/DocumentControllerTest.php
@@ -9,6 +9,7 @@ use App\Entity\Report\Document;
 use App\Entity\Report\Report;
 use App\Entity\Report\ReportSubmission;
 use App\Repository\DocumentRepository;
+use app\tests\Integration\Controller\AbstractTestController;
 
 class DocumentControllerTest extends AbstractTestController
 {
