+++ api/app/tests/Integration/Controller/NoteControllerTest.php
@@ -5,6 +5,7 @@ namespace App\Tests\Unit\Controller;
 use App\Entity\Client;
 use App\Entity\Note;
 use App\Entity\User;
+use app\tests\Integration\Controller\AbstractTestController;
 
 class NoteControllerTest extends AbstractTestController
 {
