+++ api/app/tests/Integration/Controller/PreRegistrationControllerTest.php
@@ -4,7 +4,8 @@ namespace App\Tests\Unit\Controller;
 
 use App\Entity\PreRegistration;
 use App\Entity\User;
-use App\Tests\Unit\Fixtures;
+use app\tests\Integration\Controller\AbstractTestController;
+use app\tests\Integration\Fixtures;
 
 class PreRegistrationControllerTest extends AbstractTestController
 {
