+++ api/app/tests/Integration/Controller/SelfRegisterControllerTest.php
@@ -5,6 +5,7 @@ namespace App\Tests\Unit\Controller;
 use App\Entity\Client;
 use App\Entity\PreRegistration;
 use App\Entity\User;
+use app\tests\Integration\Controller\AbstractTestController;
 
 class SelfRegisterControllerTest extends AbstractTestController
 {
