+++ api/app/tests/Integration/Controller/StatsControllerTest.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace App\Tests\Unit\Controller;
 
+use app\tests\Integration\Controller\AbstractTestController;
+
 class StatsControllerTest extends AbstractTestController
 {
     private $entityManager;
