+++ api/app/tests/Integration/ControllerReport/DecisionControllerTest.php
@@ -3,7 +3,7 @@
 namespace App\Tests\Unit\ControllerReport;
 
 use App\Entity\Report\Report;
-use App\Tests\Unit\Controller\AbstractTestController;
+use app\tests\Integration\Controller\AbstractTestController;
 
 class DecisionControllerTest extends AbstractTestController
 {
