+++ api/app/tests/Integration/ControllerReport/MoneyTransferControllerTest.php
@@ -4,7 +4,7 @@ namespace App\Tests\Unit\ControllerReport;
 
 use App\Entity\Report\MoneyTransfer;
 use App\Entity\Report\Report;
-use App\Tests\Unit\Controller\AbstractTestController;
+use app\tests\Integration\Controller\AbstractTestController;
 
 class MoneyTransferControllerTest extends AbstractTestController
 {
