+++ api/app/tests/Integration/ControllerReport/ReportControllerTest.php
@@ -7,7 +7,7 @@ use App\Entity\Report\ChecklistInformation;
 use App\Entity\Report\Document;
 use App\Entity\Report\Fee;
 use App\Entity\Report\Report;
-use App\Tests\Unit\Controller\AbstractTestController;
+use app\tests\Integration\Controller\AbstractTestController;
 
 class ReportControllerTest extends AbstractTestController
 {
