+++ api/app/tests/Integration/ControllerReport/VisitsCareControllerTest.php
@@ -3,7 +3,7 @@
 namespace App\Tests\Unit\ControllerReport;
 
 use App\Entity\Report\Report;
-use App\Tests\Unit\Controller\AbstractTestController;
+use app\tests\Integration\Controller\AbstractTestController;
 
 class VisitsCareControllerTest extends AbstractTestController
 {
