+++ api/app/tests/Integration/v2/Controller/OrganisationControllerTest.php
@@ -4,7 +4,7 @@ namespace App\Tests\Unit\v2\Controller;
 
 use App\Entity\Organisation;
 use App\Entity\User;
-use App\Tests\Unit\Controller\AbstractTestController;
+use app\tests\Integration\Controller\AbstractTestController;
 use Doctrine\ORM\EntityManager;
 use Symfony\Component\HttpFoundation\Response;
